feat(ddm): Implement transformer API
