+++ src/sentry/sentry_metrics/querying/data_v2/api.py
@@ -1,4 +1,5 @@
 from collections.abc import Mapping, Sequence
+from dataclasses import dataclass
 from datetime import datetime, timedelta, timezone
 from typing import Any
 
@@ -8,7 +9,7 @@ from sentry import features
 from sentry.models.environment import Environment
 from sentry.models.organization import Organization
 from sentry.models.project import Project
-from sentry.sentry_metrics.querying.data_v2.execution import QueryExecutor
+from sentry.sentry_metrics.querying.data_v2.execution import QueryExecutor, QueryResult
 from sentry.sentry_metrics.querying.data_v2.parsing import QueryParser
 from sentry.sentry_metrics.querying.data_v2.plan import MetricsQueriesPlan
 from sentry.sentry_metrics.querying.data_v2.preparation import (
@@ -18,6 +19,11 @@ from sentry.sentry_metrics.querying.data_v2.preparation import (
 )
 from sentry.sentry_metrics.querying.data_v2.transformation import QueryTransformer
 from sentry.utils import metrics
+
+
+@dataclass(frozen=True)
+class QueryResults:
+    results: Sequence[QueryResult]
 
 
 def _time_equal_within_bound(time_1: datetime, time_2: datetime, bound: timedelta) -> bool:
