+++ src/sentry/sentry_metrics/querying/data_v2/execution.py
@@ -11,11 +11,10 @@ from snuba_sdk.conditions import BooleanCondition, BooleanOp, Condition, Op
 from sentry.models.organization import Organization
 from sentry.models.project import Project
 from sentry.sentry_metrics.querying.common import SNUBA_QUERY_LIMIT
-from sentry.sentry_metrics.querying.data_v2.plan import QueryOrder
 from sentry.sentry_metrics.querying.data_v2.preparation import IntermediateQuery
 from sentry.sentry_metrics.querying.data_v2.units import MeasurementUnit, UnitFamily
 from sentry.sentry_metrics.querying.errors import MetricsQueryExecutionError
-from sentry.sentry_metrics.querying.types import GroupKey, GroupsCollection
+from sentry.sentry_metrics.querying.types import GroupKey, GroupsCollection, QueryOrder
 from sentry.sentry_metrics.querying.visitors import (
     QueriedMetricsVisitor,
     TimeseriesConditionInjectionVisitor,
